pyproject.toml
[project]
name = "agentic_template"
version = "0.1.0"
description = "Agentic RAG chatbot template built with Streamlit. MIT License."
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    { name = "Keming He", email = "keminghe.career@gmail.com" }
]
keywords = ["agentic", "rag", "chatbot", "streamlit", "template"]
classifiers = [
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
  "streamlit (>=1.42.0,<2.0.0)",
  "python-dotenv (>=1.0.1,<2.0.0)",
  "langchain (>=0.3.18,<0.4.0)",
  "langchain-openai (>=0.3.4,<0.4.0)",
  "langchain-pinecone (>=0.2.2,<0.3.0)",
  "pinecone (>=5.4.2,<6.0.0)",
  "langgraph (>=0.2.70,<0.3.0)",
  "langsmith (>=0.3.8,<0.4.0)",
  "duckduckgo-search (>=7.3.2,<8.0.0)"
]

[project.urls]
homepage = "https://agentic-template.streamlit.app"
repository = "https://github.com/KemingHe/agentic-template"
documentation = "https://github.com/KemingHe/agentic-template/blob/main/README.md"
"Bug Tracker" = "https://github.com/KemingHe/agentic-template/issues"

[project.scripts]
start = "streamlit run app/homepage.py"

[tool.poetry]
package-mode = false
requires-poetry = ">2.0"

[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
